+++ libs/api/database-manager/service/feature/src/lib/database-manager.service.ts
@@ -229,4 +229,64 @@ export class DatabaseManagerService {
     }
     return object;
   }
+  async changeDownloaded(id: string) {
+    const url =
+      'https://data.mongodb-api.com/app/data-dtzbr/endpoint/data/v1/action/';
+      const config = {
+        method: 'post',
+        headers: {
+          'Content-Type': 'application/json',
+          'Access-Control-Request-Headers': '*',
+          'api-key': GlobalKey.key,
+        },
+      };
+    let action = 'findOne';
+    let data = JSON.stringify({
+      collection: 'PDF',
+        database: 'PDF',
+        dataSource: 'Cluster0',
+      filter: { id: id },
+    });
+
+    const res = await lastValueFrom(
+      this.httpService.post(url + action, data, config).pipe(
+        tap((res) => console.log(res.status)),
+        map((res) => res.data)
+      )
+    );
+
+    console.log(res);
+
+    action = 'updateOne';
+    data = JSON.stringify({
+      collection: 'PDF',
+        database: 'PDF',
+        dataSource: 'Cluster0',
+      filter: { id: id },
+      update: {
+        $set: { downloaded: !res.document.downloaded },
+      },
+    });
+
+    //Updates the name
+    this.httpService.post(url + action, data, config).pipe(
+      tap((res) => console.log(res.status)),
+      map((res) => res.data)
+    );
+
+    //return updated record
+    action = 'findOne';
+    data = JSON.stringify({
+      collection: 'PDF',
+        database: 'PDF',
+        dataSource: 'Cluster0',
+      filter: { id: id },
+    });
+    return await lastValueFrom(
+      this.httpService.post(url + action, data, config).pipe(
+        tap((res) => console.log(res.status)),
+        map((res) => res.data)
+      )
+    );
+  }
 }
