+++ libs/api/pdf-manager/service/src/lib/api-pdf-manager-service.module.ts
@@ -1,13 +1,13 @@
 import { Module } from '@nestjs/common';
 import { ApiPdfManagerServiceService } from './api-pdf-manager-service.service';
-// import { HttpModule } from '@nestjs/axios';
+import { HttpModule } from '@nestjs/axios';
 import { CqrsModule } from '@nestjs/cqrs';
 import * as CommandHandlers from './commands/handlers';
 import { PdfManagerRepositoryDataAccessModule } from '@conversation-catcher/api/pdf-manager/repository/data-access';
 
 @Module({
   controllers: [],
-  imports: [CqrsModule, PdfManagerRepositoryDataAccessModule],
+  imports: [CqrsModule, PdfManagerRepositoryDataAccessModule, HttpModule],
   providers: [ApiPdfManagerServiceService, CommandHandlers.SetDownloadedPdfHandler],
   exports: [ApiPdfManagerServiceService],
 })
