+++ libs/api/shell/src/lib/api-shell.module.ts
@@ -14,7 +14,7 @@ import { ApiStudentExploreApiFeatureModule as PdfManager } from '@conversation-c
       driver: ApolloDriver,
       playground: true,
     }),
-    // DatabaseManager,
+    DatabaseManager,
     PdfManager,
   ],
 })
